move methods to jobService
