+++ src/main/java/uk/gov/ons/fwmt/job_service_v2/queuereceiver/JobServiceMessageReceiver.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.ons.fwmt.fwmtgatewaycommon.data.FWMTCancelJobRequest;
 import uk.gov.ons.fwmt.fwmtgatewaycommon.data.FWMTCreateJobRequest;
-import uk.gov.ons.fwmt.job_service_v2.service.tm.JobService;
+import uk.gov.ons.fwmt.job_service_v2.service.JobService;
 
 import javax.xml.datatype.DatatypeConfigurationException;
 import java.io.IOException;
