+++ src/main/java/uk/gov/ons/fwmt/job_service_v2/service/JobService.java
@@ -1,5 +1,6 @@
-package uk.gov.ons.fwmt.job_service_v2.service.tm;
+package uk.gov.ons.fwmt.job_service_v2.service;
 
+import uk.gov.ons.fwmt.fwmtgatewaycommon.data.DummyTMResponse;
 import uk.gov.ons.fwmt.fwmtgatewaycommon.data.FWMTCancelJobRequest;
 import uk.gov.ons.fwmt.fwmtgatewaycommon.data.FWMTCreateJobRequest;
 
@@ -9,4 +10,6 @@ public interface JobService {
   void createJob(FWMTCreateJobRequest jobRequest) throws DatatypeConfigurationException;
 
   void cancelJob(FWMTCancelJobRequest cancelRequest);
+
+  void notifyRM(DummyTMResponse dummyTMResponse);
 }
